Move tailwind config to plugin to allow downstream repositories to better install/use utilities
